Refatoração da camada de models Criar uma nova camada de serviço separada da camada models a fim de organizar melhor o código
